+++ brian2/codegen/runtime/weave_rt/templates/ratemonitor.cpp
@@ -3,14 +3,14 @@
     {{ common.insert_group_preamble() }}
     {# USES_VARIABLES { N, t, rate, _clock_t, _clock_dt, _spikespace,
                         _num_source_neurons, _source_start, _source_stop } #}
-	int _num_spikes = {{_spikespace}}[_num_spikespace-1];
+    size_t _num_spikes = {{_spikespace}}[_num_spikespace-1];
     // For subgroups, we do not want to record all spikes
     // We assume that spikes are ordered
     int _start_idx = -1;
     int _end_idx = - 1;
-    for(int _j=0; _j<_num_spikes; _j++)
+    for(size_t _j=0; _j<_num_spikes; _j++)
     {
-        const int _idx = {{_spikespace}}[_j];
+        const size_t _idx = {{_spikespace}}[_j];
         if (_idx >= _source_start) {
             _start_idx = _j;
             break;
@@ -18,9 +18,9 @@
     }
     if (_start_idx == -1)
         _start_idx = _num_spikes;
-    for(int _j=_start_idx; _j<_num_spikes; _j++)
+    for(size_t _j=_start_idx; _j<_num_spikes; _j++)
     {
-        const int _idx = {{_spikespace}}[_j];
+        const size_t _idx = {{_spikespace}}[_j];
         if (_idx >= _source_stop) {
             _end_idx = _j;
             break;
