+++ brian2/codegen/runtime/weave_rt/templates/spikemonitor.cpp
@@ -7,24 +7,24 @@
     {#  Get the name of the array that stores these events (e.g. the spikespace array) #}
     {% set _eventspace = get_array_name(eventspace_variable) %}
 
-    int _num_events = {{_eventspace}}[_num{{eventspace_variable.name}}-1];
+    size_t _num_events = {{_eventspace}}[_num{{eventspace_variable.name}}-1];
     if (_num_events > 0)
     {
         // For subgroups, we do not want to record all spikes
         // We assume that spikes are ordered
-        int _start_idx = _num_events;
-        int _end_idx = _num_events;
-        for(int _j=0; _j<_num_events; _j++)
+        size_t _start_idx = _num_events;
+        size_t _end_idx = _num_events;
+        for(size_t _j=0; _j<_num_events; _j++)
         {
-            const int _idx = {{_eventspace}}[_j];
+            const size_t _idx = {{_eventspace}}[_j];
             if (_idx >= _source_start) {
                 _start_idx = _j;
                 break;
             }
         }
-        for(int _j=_num_events-1; _j>=_start_idx; _j--)
+        for(size_t _j=_num_events-1; _j>=_start_idx; _j--)
         {
-            const int _idx = {{_eventspace}}[_j];
+            const size_t _idx = {{_eventspace}}[_j];
             if (_idx < _source_stop) {
                 break;
             }
@@ -46,10 +46,10 @@
             {{c_data_type(var.dtype)}}* _{{varname}}_data = ({{c_data_type(var.dtype)}}*)(((PyArrayObject*)(PyObject*){{get_array_name(var, access_data=False)}}.attr("data"))->data);
             {% endfor %}
             // Copy the values across
-            for(int _j=_start_idx; _j<_end_idx; _j++)
+            for(size_t _j=_start_idx; _j<_end_idx; _j++)
             {
-                const int _idx = {{_eventspace}}[_j];
-                const int _vectorisation_idx = _idx;
+                const size_t _idx = {{_eventspace}}[_j];
+                const size_t _vectorisation_idx = _idx;
                 {{vector_code|autoindent}}
                 {% for varname in record_variables | sort%}
                 _{{varname}}_data[_curlen + _j - _start_idx] = _to_record_{{varname}};
