+++ brian2/devices/cpp_standalone/templates/group_variable_set.cpp
@@ -4,15 +4,15 @@
 	{# USES_VARIABLES { _group_idx } #}
 	//// MAIN CODE ////////////
 	// scalar code
-    const int _vectorisation_idx = -1;
+    const size_t _vectorisation_idx = -1;
     {{scalar_code|autoindent}}
 
     {{ openmp_pragma('parallel-static') }}
-	for(int _idx_group_idx=0; _idx_group_idx<_num_group_idx; _idx_group_idx++)
+	for(int _idx_group_idx=0; _idx_group_idx<(int)_num_group_idx; _idx_group_idx++)
 	{
 	    // vector code
-		const int _idx = {{_group_idx}}[_idx_group_idx];
-		const int _vectorisation_idx = _idx;
+		const size_t _idx = {{_group_idx}}[_idx_group_idx];
+		const size_t _vectorisation_idx = _idx;
         {{vector_code|autoindent}}
 	}
 {% endblock %}
